renamed "supervized" to "supervised" to please IdoZ
