+++ synchronization/supervised/doc.go
@@ -1,5 +1,5 @@
-// Package supervized provides basic supervision abilities for running goroutines,
+// Package supervised provides basic supervision abilities for running goroutines,
 // namely making sure that panics are not swallowed and that long-running goroutines
 // are restarted if they crash
-package supervized
+package supervised
